backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Havshyn Maksym",email = "maksgavshin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.123.4,<0.124.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "password-strength (>=0.0.3.post2,<0.0.4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
